+++ clients/fides-js/src/fides-tcf.ts
@@ -221,7 +221,6 @@ const _Fides: FidesGlobal = {
     allowHTMLDescription: null,
     base64Cookie: false,
     fidesPrimaryColor: null,
-    forceGpp: false,
     fidesClearCookie: false,
   },
   fides_meta: {},